Move LUT to utility module
